Add support for selectively ignoring system call errors
While using this tool as part of a larger configuration management
system based on Puppet, it was found necessary to ignore certain types
of errors. In this particular case, the ENOENT (No such file or
directory) error code needed to be ignored, so that Puppet will create
the directory if it is not present.

The bash-like pseudocode describing this situation is:

if nfs-mountpoint-check --ignore-errno=2 /mnt/point ; then
	# make sure /mnt/point directory exists
	# make sure /mnt/point directory is mounted
else
	# force lazy umount, or similar recovery action
fi

Without the ability to ignore ENOENT, it is not possible to write this
style of code. The usual test to check if the /mnt/point directory
exists uses the stat() system call. This system call can hang forever if
the NFS server is not responding.