Add Large File Support
When accessing an NFS server which runs a 64 bit OS and exports a 64 bit
filesystem (for example, a multi-TiB XFS filesystem) the NFS clients
will be presented with 64 bit inode numbers.

If the NFS client is a 32 bit client, then glibc will return an
EOVERFLOW (errno = 75) error code from functions such as fstat() which
return the inode number as part of their result value. This is because
the 64 bit inode numbers are too large to fit into the 32 bit inode
field on these machines. (For example, see the "struct stat" "st_ino"
field.)

To avoid this problem, we build with Large File Support. Adding these
options to our compiler command line instructs glibc to transparently
use 64 bit variants of fstat() (and related functions), even on 32 bit
systems.